refactor: 기존 토스페이먼츠 결제 승인 api 함수의 타입 별도로 분리 - 타입 분리에 따른 기존 코드의 경로 업데이트
